move price preds add hasPrices hasPriceModifiers
