+++ packages/predicates/src/common/index.ts
@@ -4,6 +4,5 @@ export * from './isExpired';
 export * from './isTrashed';
 export * from './isValidOrTrashed';
 export * from './selectionById';
-export * from './selectionPredicates';
 export * from './listPredicates';
 export * from './misc';
