+++ packages/predicates/src/priceTypes/selectionPredicates/index.ts
@@ -1,7 +1,8 @@
 import { allPass, find } from 'ramda';
 
+import { isNotBasePrice, isNotDiscount, isNotPercent } from '../../prices';
+
 import type { PriceType } from '@eventespresso/edtr-services';
-import { isNotBasePrice, isNotDiscount, isNotPercent } from '../../common';
 
 // returns true if supplied price type is a flat fee (dollar) surcharge
 export const isFlatFeeSurcharge = allPass([isNotBasePrice, isNotDiscount, isNotPercent]);
